move MKL implementation to dedicated TPL header
