+++ src/blas/impl/KokkosBlas_util.hpp
@@ -47,15 +47,6 @@
 
 #include "Kokkos_ArithTraits.hpp"
 
-// TPL macros
-#if defined(KOKKOSKERNELS_ENABLE_TPL_MKL)
-#include "mkl_version.h"
-#if __INTEL_MKL__ >= 2018
-#define __KOKKOSBLAS_ENABLE_INTEL_MKL_COMPACT__ 1
-#include "mkl_compact.h"
-#endif
-#endif
-
 namespace KokkosBlas {
 
 //////// Tags for BLAS ////////
